utils: move batch utils to different source file
