+++ src/util/func.cpp
@@ -47,30 +47,6 @@ std::string buildAsyncResponse(const faabric::Message& msg)
     return std::to_string(msg.id());
 }
 
-std::shared_ptr<faabric::BatchExecuteRequest> batchExecFactory()
-{
-    auto req = std::make_shared<faabric::BatchExecuteRequest>();
-    req->set_id(faabric::util::generateGid());
-    return req;
-}
-
-std::shared_ptr<faabric::BatchExecuteRequest> batchExecFactory(
-  const std::string& user,
-  const std::string& function,
-  int count)
-{
-    auto req = batchExecFactory();
-
-    // Force the messages to have the same app ID
-    uint32_t appId = faabric::util::generateGid();
-    for (int i = 0; i < count; i++) {
-        *req->add_messages() = messageFactory(user, function);
-        req->mutable_messages()->at(i).set_appid(appId);
-    }
-
-    return req;
-}
-
 std::shared_ptr<faabric::Message> messageFactoryShared(
   const std::string& user,
   const std::string& function)
